# api/openapi.yaml
openapi: 3.1.0
info:
  title: certctl API
  description: |
    Certificate lifecycle management platform API. Manages certificates, issuers,
    deployment targets, agents, jobs, policies, profiles, teams, owners, agent groups,
    audit events, notifications, and observability metrics.

    All endpoints under `/api/v1/` require authentication by default (configurable via
    `CERTCTL_AUTH_TYPE`). Use `Bearer {api_key}` in the Authorization header.

    Paginated list endpoints accept `page` (default 1) and `per_page` (default 50, max 500)
    query parameters and return a standard envelope with `data`, `total`, `page`, and `per_page`.
  # ARCH-001-A closure (Sprint 5, 2026-05-16): info.version MUST track
  # the latest `v*` git tag. The openapi-version-tag-parity.sh CI guard
  # asserts this on every CI run. Bump in lockstep with the
  # `git tag -a v* ...` command at release time.
  version: 2.1.7
  license:
    name: BSL 1.1
    url: https://github.com/zulufun/certctl-localhost/blob/master/LICENSE
servers:
  - url: https://localhost:8443
    description: Docker Compose demo (self-signed cert; pin with ./deploy/test/certs/ca.crt)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/audit/export:
    get:
      tags: [Audit]
      summary: Export audit events as newline-delimited JSON (NDJSON) for a date range
      description: |
        Permission `audit.export`. Streams every audit row inside the
        requested `[from, to]` window as `application/x-ndjson`. Used
        by compliance pipelines (Splunk Universal Forwarder, Elastic
        Filebeat, Vector, etc.) that prefer line-by-line ingestion
        over a single JSON document.

        Range cap: 90 days. Requests with `to - from > 90d` return
        400; paginate by narrower windows.

        Per-record cap: `limit` query parameter (default 50000;
        accepted range 1..100000). Values outside the range silently
        clamp to default.

        The export itself is recursively audited: every successful
        export emits an `audit.export` event capturing actor, range,
        category, and row count so the audit log records who pulled
        which compliance evidence and when.
      operationId: exportAudit
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date-time }
          description: RFC 3339 start of the export window (inclusive).
        - in: query
          name: to
          required: true
          schema: { type: string, format: date-time }
          description: RFC 3339 end of the export window (exclusive). Must be strictly after `from`.
        - in: query
          name: category
          required: false
          schema:
            type: string
            enum: [cert_lifecycle, auth, config]
          description: Optional category filter. Omit to return every event in the window.
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100000 }
          description: Maximum rows to stream (default 50000; out-of-range values clamp to default).
      responses:
        "200":
          description: NDJSON stream of AuditEvent rows
          headers:
            Content-Disposition:
              description: Attachment hint suggesting filename `certctl-audit-<from>_to_<to>.ndjson`.
              schema: { type: string }
          content:
            application/x-ndjson:
              schema:
                type: string
                description: |
                  One JSON-encoded AuditEvent per line (see
                  `#/components/schemas/AuditEvent`). The body is a
                  byte stream, not a single JSON document; clients
                  parse line-by-line.
        "400": { description: Missing/invalid date params, range exceeds 90 days, invalid category, etc. }
        "401": { description: Unauthorized }
        "403": { description: Forbidden }
        "405": { description: Method not allowed (only GET is accepted) }
        "500": { description: Internal error (export failed mid-query) }
...
  /api/v1/certificates:
    get:
      tags: [Certificates]
      summary: List certificates
      operationId: listCertificates
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/per_page"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CertificateStatus"
        - name: environment
          in: query
          schema:
            type: string
        - name: owner_id
          in: query
          schema:
            type: string
        - name: team_id
          in: query
          schema:
            type: string
        - name: issuer_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of certificates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ManagedCertificate"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Certificates]
      summary: Create certificate
      operationId: createCertificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagedCertificate"
      responses:
        "201":
          description: Certificate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedCertificate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/certificates/{id}/export/pem:
    get:
      tags: [Certificates]
      summary: Export certificate as PEM
      description: |
        Returns the certificate and its chain in PEM format. By default returns JSON
        with cert_pem, chain_pem, and full_pem fields. Add ?download=true to get the
        full PEM chain as a file download with Content-Disposition headers.
      operationId: exportCertificatePEM
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - name: download
          in: query
          schema:
            type: string
            enum: ["true"]
          description: Set to "true" to get a file download instead of JSON.
      responses:
        "200":
          description: PEM export
          content:
            application/json:
              schema:
                type: object
                properties:
                  cert_pem:
                    type: string
                    description: Leaf certificate PEM
                  chain_pem:
                    type: string
                    description: Intermediate/root chain PEM
                  full_pem:
                    type: string
                    description: Full PEM chain (cert + intermediates)
            application/x-pem-file:
              schema:
                type: string
                format: binary
                description: Full PEM file (when download=true)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/certificates/{id}/export/pkcs12:
    post:
      tags: [Certificates]
      summary: Export certificate as PKCS#12
      description: |
        Returns a PKCS#12 (.p12) bundle containing the certificate and chain.
        Private keys are NOT included — they live on agents and never touch the control plane.
        The bundle is encrypted with the provided password (or empty password if omitted).
      operationId: exportCertificatePKCS12
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: Password to encrypt the PKCS#12 bundle (can be empty)
      responses:
        "200":
          description: PKCS#12 binary
          content:
            application/x-pkcs12:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── PKI (CRL & OCSP, RFC 5280 / 6960 / 8615) ──────────────────────
  #
  # Relying parties (browsers, OpenSSL clients, OCSP stapling sidecars,
  # mTLS clients) cannot present a certctl Bearer token, so these two
  # endpoints are unauthenticated and live under the RFC 8615
  # `.well-known` namespace. They were previously mounted at
  # /api/v1/crl/{issuer_id} and /api/v1/ocsp/{issuer_id}/{serial}; those
  # paths were removed in M-006.
  #
  # The non-standard JSON CRL endpoint (GET /api/v1/crl) was also
  # removed — RFC 5280 defines only the DER wire format.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key passed as Bearer token. Configure via CERTCTL_AUTH_SECRET.
    # Auth Bundle 2 Phase 5 — session-cookie auth scheme. New
    # session-authenticated endpoints declare
    # `security: [{cookieAuth: []}, {bearerAuth: []}]` (either auth
    # method works, OR semantics). Per Phase 5 spec, the
    # `/auth/oidc/back-channel-logout` endpoint declares `security: []`
    # because auth comes from the IdP-signed logout token in the body,
    # not certctl-issued credentials.
    cookieAuth:
      type: apiKey
      in: cookie
      name: certctl_session
      description: |
        Session cookie minted by `POST /auth/oidc/callback` after a
        successful OIDC handshake (Auth Bundle 2). Wire format
        `v1.<session_id>.<signing_key_id>.<HMAC-SHA256>`; HMAC is
        verified server-side against the active session signing key.
        Cookie attributes: `Secure` `HttpOnly` `SameSite=Lax|Strict`
        (configurable via `CERTCTL_SESSION_SAMESITE`) `Path=/`.
        State-changing requests additionally require the
        `X-CSRF-Token` header to match the SHA-256 hash on the
        session row (validated by the session middleware in Phase 6).
  parameters:
    resourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Human-readable resource ID (e.g., mc-api-prod, t-platform)
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    per_page:
      name: per_page
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        request_id:
          type: string

    # ARCH-001-A closure (Sprint 5, 2026-05-16). Three operation
    # responses (search `#/components/schemas/Error` in this file)
    # reference a schema named "Error" — but only "ErrorResponse" was
    # defined, so the orval codegen failed with
    # MissingPointerError. Alias Error → ErrorResponse so the spec
    # parses cleanly and the three offenders keep their stable
    # response shape.
...
    PaginationEnvelope:
      type: object
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
        per_page:
          type: integer

    # ─── Certificates ────────────────────────────────────────────────
    CertificateStatus:
      type: string
      enum:
        - Pending
        - Active
        - Expiring
        - Expired
        - RenewalInProgress
        - Failed
        - Revoked
        - Archived
    ManagedCertificate:
      # D-5 (cat-f-ae0d06b6588f, master): per-issuance fields
      # (serial_number, fingerprint_sha256, key_algorithm, key_size,
      # issued_at) are intentionally NOT declared here. They live on
      # CertificateVersion (per-issuance evidence) and are fetched via
      # /api/v1/certificates/{id}/versions. ManagedCertificate is the
      # management envelope; CertificateVersion is the issuance record.
      # Pre-D-5 the TS Certificate interface had them as optional and
      # the dashboard's Key Algorithm / Key Size rows always rendered
      # '—' as a result. The TS trim restores parity with this schema.
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        common_name:
          type: string
        sans:
          type: array
          items:
            type: string
        environment:
          type: string
        owner_id:
          type: string
        team_id:
          type: string
        issuer_id:
          type: string
        target_ids:
          type: array
          items:
            type: string
        renewal_policy_id:
          type: string
        certificate_profile_id:
          type: string
        status:
          $ref: "#/components/schemas/CertificateStatus"
        expires_at:
          type: string
          format: date-time
        tags:
          type: object
          additionalProperties:
            type: string
        last_renewal_at:
          type: string
          format: date-time
        last_deployment_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        revocation_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
        - common_name
        - renewal_policy_id
        - issuer_id
        - owner_id
        - team_id
...